diarization/requirements.txt
numpy<2.0
huggingface_hub==0.23.4
pyannote.audio==3.1.1
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
torch==2.1.2
torchaudio==2.1.2
